CATALOG-2925: Change to new catalog
