+++ query/src/catalogs/impls/catalog/mod.rs
@@ -15,11 +15,12 @@
 
 mod immutable_catalog;
 mod mutable_catalog;
+mod overlaid_catalog1;
 
 pub mod metastore_catalog;
 pub mod overlaid_catalog;
-mod overlaid_catalog1;
 pub mod system_catalog;
 
 pub use immutable_catalog::ImmutableCatalog;
 pub use mutable_catalog::MutableCatalog;
+pub use overlaid_catalog1::OverlaidCatalog1;
