+++ query/src/catalogs/impls/mod.rs
@@ -17,7 +17,9 @@ mod catalog;
 mod database_catalog;
 pub mod in_memory_meta;
 
+pub use catalog::ImmutableCatalog;
 pub use catalog::MutableCatalog;
+pub use catalog::OverlaidCatalog1;
 pub use database_catalog::DatabaseCatalog;
 
 pub use crate::catalogs::table_id_ranges::LOCAL_TBL_ID_BEGIN;
