+++ query/src/catalogs/mod.rs
@@ -23,6 +23,7 @@ pub mod backends;
 pub mod impls;
 
 pub use catalog::Catalog;
+pub use catalog::Catalog1;
 pub use database::Database;
 pub use database::Database1;
 pub use impls::in_memory_meta::InMemoryMetas;
