+++ query/src/interpreters/interpreter_database_drop.rs
@@ -19,7 +19,7 @@ use common_planners::DropDatabasePlan;
 use common_streams::DataBlockStream;
 use common_streams::SendableDataBlockStream;
 
-use crate::catalogs::Catalog;
+use crate::catalogs::Catalog1;
 use crate::interpreters::Interpreter;
 use crate::interpreters::InterpreterPtr;
 use crate::sessions::DatabendQueryContextRef;
