+++ query/src/sessions/context_shared.rs
@@ -30,7 +30,6 @@ use futures::future::AbortHandle;
 use uuid::Uuid;
 
 use crate::catalogs::impls::DatabaseCatalog;
-use crate::catalogs::Catalog;
 use crate::catalogs::Table;
 use crate::clusters::ClusterRef;
 use crate::configs::Config;
