+++ query/src/sql/statements/query/query_schema_joined_analyzer.rs
@@ -23,7 +23,7 @@ use sqlparser::ast::TableAlias;
 use sqlparser::ast::TableFactor;
 use sqlparser::ast::TableWithJoins;
 
-use crate::catalogs::Catalog;
+use crate::catalogs::Catalog1;
 use crate::sessions::DatabendQueryContextRef;
 use crate::sql::statements::analyzer_expr::ExpressionAnalyzer;
 use crate::sql::statements::query::query_schema_joined::JoinedSchema;
